Trying to implement register. how register work is by activating user grain and check if the password is null or not. If not null, that means this user grain with key username already exist
current status of the work, in short, not yet dont.
for longer info it can safely persist user, but somehow cannot access
another grains from a grain, to be specific, I want to communicate with password
hasher grains but failed to do so. And even thou the process is done, the return
value for the client is not.