removing direct access db from Conduit project. Npgsql in Conduit project now used only for accessing cluster management table, where previousnya it was used for getting username from email.
The way forward, there's a grain to handle email to username transformation.

We decided the new grain, EmailUserGrain, is tightly coupled with UserGrain. When UserGrain is activated, either registering user or changing profile, EmailUserGrain will also be activated. Its safe to say EmailUserGrain work like somekind of cache but will be updated if the source of truth (user state) is updated.