Fix bug counter not working.
CounterState was created with immutability in mind, so the only field of the CounterState class, which is Current, was being made readonly. Two constructor was given, the one without parameters so it will set the Current to zero, and the other one with param of the value. Turns out, Orleans cannot invoke the constructor with parameters, so CounterState was being made mutable.